requirements.txt
# requirements.txt
Flask
stripe
gspread
oauth2client
python-dotenv
Flask-Cors
